Support dynamic import examples for C/C++
