+++ examples/C_Cpp/examples/logging.conf
@@ -1,7 +1,7 @@
 ####################### Loggers #######################
 
 [loggers]
-keys=root,tlkcore,comm,device,calitable,aakittable,beamtable,udtable,db
+keys=root,tlkcore,comm,device,calitable,aakittable,beamtable,udtable,db,dfu
 
 # Default logger
 [logger_root]
@@ -49,6 +49,11 @@ propagate=0
 handlers=libDetailfileHandler,libconsoleHandler
 #handlers=libDetailfileHandler
 qualname=Device
+propagate=0
+
+[logger_dfu]
+handlers=libfileHandler,consoleHandler
+qualname=DFU
 propagate=0
 
 [logger_beamconfig]
